Created the favourite base class
